feat(setup): Finished proyect set up using atomic design principle
