+++ .gitignore
@@ -1,0 +1,37 @@
+# Archivos de log
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+# Directorio de dependencias. ¡Nunca subas esto!
+node_modules/
+
+# Directorio de build de Vite
+dist/
+dist-ssr/
+
+# Archivos de entorno
+# Contienen variables de entorno y secretos.
+# Crea un archivo .env.example para que otros sepan qué variables se necesitan.
+.env
+.env.local
+.env.*.local
+
+# Archivos de caché y temporales
+.npm
+.yarn-integrity
+.pnp.cjs
+.pnp.loader.mjs
+
+# Archivos de configuración del editor
+.idea/
+.vscode/*
+!.vscode/extensions.json
+
+# Archivos generados por el sistema operativo
+.DS_Store
+Thumbs.db
