--- landing.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-  <meta charset="UTF-8">
-  <title>E-Commerce | Landing</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <header>
-    <h1>Tienda Online</h1>
-    <nav>
-      <a href="#">Inicio</a>
-      <a href="#">Catálogo</a>
-      <a href="#">Login</a>
-    </nav>
-  </header>
-
-  <main>
-    <section class="bienvenida">
-      <h2>¡Bienvenido a nuestra tienda!</h2>
-      <p>Explorá nuestros productos y hacé tu compra de forma fácil y rápida.</p>
-    </section>
-
-    <section class="destacados">
-      <h3>Productos Destacados</h3>
-      <div class="productos">
-        <div class="producto">
-          <img src="https://picsum.photos/200" alt="Producto aleatorio">
-          <p>Producto 1</p>
-        </div>
-        <div class="producto">
-          <img src="https://picsum.photos/202" alt="Producto aleatorio">
-          <p>Producto 2</p>
-        </div>
-      </div>
-    </section>
-  </main>
-
-  <footer>
-    <p>© 2025 Tienda Online. Todos los derechos reservados.</p>
-  </footer>
-</body>
-</html>
