+++ src/App.jsx
@@ -1,0 +1,11 @@
+import Home from './components/views/Home/Home';
+
+function App() {
+  return (
+    <>
+      <Home />
+    </>
+  );
+}
+
+export default App;
