+++ src/components/atoms/atoms.md
@@ -1,0 +1,3 @@
+# Atomos (Los componentes mas pequenios)
+
+## Ejemplo: Botón, input, icono
