+++ src/components/molecules/molecules.md
@@ -1,0 +1,3 @@
+# Moleculas (Compuestos por atomos)
+
+## Ejemplo: Formulario de login (con inputs y botón)
