+++ src/components/organisms/Carrousel/Carrousel.css
