+++ src/components/organisms/Header/Header.css
@@ -28,6 +28,7 @@
   text-decoration: underline;
 }
 
+
 .buscador {
   margin-top: 1rem;
   display: flex;
