+++ src/components/organisms/organisms.md
@@ -1,0 +1,3 @@
+# Organismos
+
+## Ejemplo: Header, Footer, Sidebar
