+++ src/components/views/views.md
@@ -1,0 +1,3 @@
+# Views
+
+## Ejemplo: Home, Productos, Registro, Login
