--- style.css
@@ -1,34 +1,0 @@
-body {
-  font-family: sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-header {
-  background-color: #c7b3b3;
-  color: white;
-  padding: 1rem;
-  display: flex; 
-  justify-content: space-between;
-  width: 100%;
-}
-
-nav a {
-  color: white;
-  margin-left: 1rem;
-  text-decoration: none;
-}
-
-main {
-  padding: 2rem;
-}
-
-.productos {
-  display: flex;
-  gap: 1rem;
-}
-
-.producto img {
-  width: 150px;
-  height: 150px;
-}
